Rework the error responses for characters expected but missing
This will help considerably in keeping the (in-progress) OTPCL shell
nice and simple.  As a side effect, it also helps clean up
?TOKEN_TAKE_DESCEND_FLATTEN, since now it returns something actually
applicable instead of always pretending to be a curly-braced
string (granted, curly-braced things - both strings and variables -
are the only things that actually use that right now, but if that
changes in the future we're prepared).