Expose otpcl_parse:parse/2
I had envisioned this for otpcl_shell, but as I was writing the commit
message I realized that it wouldn't be a good idea (since it would
complicate things if the shell input included multiple commands,
e.g. semicolon-separated).  Figured it might be useful elsewhere,
though (e.g. in tests), so I'm committing it anyway.