(VERY WIP) first pass on otpcl_shell
This is very much a work in progress.  All I know is that it
successfully compiles.  I have no idea if it actually works.  In fact,
it probably doesn't.  It took a lot of studying of how the Erlang,
Elixir, LFE, and rebar3 shells work, but I think/hope I got the gist
of it.  We'll find out as soon as I get around to, you know, actually
trying this out and debugging it.