Add atom creation warning for the `cmd` command.
When generating the body for a pure-OTPCL command definition, the
generator produces atoms for the names of temporary state
variables (`State0`, `State1`, ...).  It's highly unlikely that this
is particularly dangerous (you'd have to have a *really* long command
body to blow out the atom ceiling this way), but it's worth flagging
nonetheless.